+++ inlang/source-code/create-project/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @inlang/create-project
 
+## 1.1.8
+
+### Patch Changes
+
+- Updated dependencies [87bed968b]
+- Updated dependencies [23ca73060]
+  - @inlang/sdk@0.25.0
+
+## 1.1.7
+
+### Patch Changes
+
+- @inlang/sdk@0.24.1
+
 ## 1.1.6
 
 ### Patch Changes