+++ inlang/source-code/ide-extension/CHANGELOG.md
@@ -1,5 +1,27 @@
 # inlang-vs-code-extension
 
+## 1.26.0
+
+### Minor Changes
+
+- 2f55a1a0d: chore: add project group identify
+
+### Patch Changes
+
+- Updated dependencies [87bed968b]
+- Updated dependencies [23ca73060]
+  - @inlang/sdk@0.25.0
+  - @inlang/create-project@1.1.8
+  - @inlang/telemetry@0.3.7
+
+## 1.25.5
+
+### Patch Changes
+
+- @inlang/sdk@0.24.1
+- @inlang/create-project@1.1.7
+- @inlang/telemetry@0.3.6
+
 ## 1.25.4
 
 ### Patch Changes