+++ inlang/source-code/plugins/json/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @inlang/plugin-json
 
+## 5.1.8
+
+### Patch Changes
+
+- Updated dependencies [87bed968b]
+- Updated dependencies [23ca73060]
+  - @inlang/sdk@0.25.0
+
+## 5.1.7
+
+### Patch Changes
+
+- @inlang/sdk@0.24.1
+
 ## 5.1.6
 
 ### Patch Changes